fix(add-remote): widen host_port to int32, trim pubkey whitespace, stub remote   status check

+++ backend/routes/remotes/check_remote.go
@@ -36,12 +36,12 @@ func CheckRemote(remoteContext *AddRemoteDTO, logger *zerolog.Logger, privateKey
 
 	conn, err := ssh.Dial("tcp", net.JoinHostPort(remoteContext.Host, strconv.Itoa(int(remoteContext.HostPort))), clientConfig)
 	if err != nil {
-		logger.Err(err).Str("remote host", remoteContext.Host).Int16("remote port", remoteContext.HostPort).Msgf("cannot reach server")
+		logger.Err(err).Str("remote host", remoteContext.Host).Int32("remote port", remoteContext.HostPort).Msgf("cannot reach server")
 		return "", err
 	}
 	defer func() {
 		if err := conn.Close(); err != nil {
-			logger.Err(err).Str("remote host", remoteContext.Host).Int16("remote port", remoteContext.HostPort).Msg("cannot close connection with server")
+			logger.Err(err).Str("remote host", remoteContext.Host).Int32("remote port", remoteContext.HostPort).Msg("cannot close connection with server")
 		}
 	}()
